requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
SpeechRecognition==3.10.0
pyttsx3==2.90
requests==2.31.0
Pillow==10.0.0 